Added Relic Fragment, Win Screen, Item System, And Main Menu
A Simple Main Menu  was added.

A new Item system was added, extending from the sign's basic functionality, so that you can obtain an item upon interacting it with, and it will display a prompt.

A Win Screen was implemented, Once you have collected all 4 relic pieces, you can choose to exit to menu or continue playing.